+++ package.json
@@ -3,7 +3,8 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "nodemon http://localhost:3000 ./bin/www"
+    "start": "node --watch app.js",
+    "dev": "nodemon http://localhost:3000 ./bin/www"
   },
   "dependencies": {
     "@prisma/client": "^5.19.1",
@@ -16,9 +17,10 @@
     "http-errors": "~1.6.3",
     "jsonwebtoken": "^9.0.2",
     "morgan": "~1.9.1",
-    "nodemon": "^3.1.4"
+    "projeto": "file:"
   },
   "devDependencies": {
+    "nodemon": "^3.1.4",
     "prisma": "^5.19.1"
   }
 }
